Added "assumes: [snapd2.55]" line to snapcraft.yaml
The "cups" interface, a slot of the CUPS Snap for client Snaps with
print functionality to plug, needs explicit support in snapd,
to bind-mount the directory with the domain socket of CUPS into
the file system of the client, to point the CUPS_SERVER environment
variable pointing to the domain socket in the mounted directory, and
to auto-install the CUPS Snap when a "cups"-plugging client Snap
is installed but the CUPS Snap not present.

All this is only implemented in snapd 2.55 or later, therefore we
require this snapd version via the "assumes: [snapd2.55]" entry.