Ghostscript: add appleraster to the driver list
cups-filter, since 056d2b5 ("libcupsfilters: For ghostscript() require
GS 9.56.0"), seems to unconditionally require GS's "appleraster" driver.
As it seems not to be built with the current driver list, add it to the
driver list.

This fixes printing to AirPrint-enabled printers - or at least, it fixes
printing to my Brother MFC J430W.