snapcraft.yaml: Plug "cups-control" for system CUPS' D-Bus service access
In proxy mode cups-proxyd has to keep the cloned queues on the Snap's
CUPS with the original queues on the system's CUPS. For this it gets
notified by DNS-SD and by D-Bus, for the latter subscribing to the
notification D-Bus service of the system's CUPS. This did not work as
the CUPS Snap needs to plug the system's "cups-control" in order to
access the D-Bus service of the system's CUPS. This we are doing now.

To be able to plug "cups-control" on the system our Snap cannot
provide a slot named "cups-control". Therefore we rename this slot to
"cups-server".

While doing this it turned out that the "cups-internal" interface
defined in the CUPS Snap is actually not needed. It was originally
intended to allow the included command line tools to access the
included CUPS daemon. This is not needed. Any part of the Snap can
always access any other, even if they are running in different
processes. So the "cups-internal" interface got removed now.

Thanks, Ken VanDine and James Henstridge, for all the hints and the
patience with me to sort that out.