Stick to commit e496badbf of cups-filters
As we are doing a major change in cups-filters (libppd depends on
libcupsfilters and not libcupsfilters depends on libppd any more) we
do not want to make the Snap Store builds grab the bleeding edge of
cups-filters, to have some time to test the changes.

This is especially important as the CUPS Snap got integral part of
snapd for its `cups` printing interface.

Marked CUPS Snap release 2.4.2-3.