Preset maximum cups-browsed_log file size to 30 MB instead of 300 kB
Debug logging of cups-browsed is very verbose and so the 300-kB
default is very quickly filled up, so that important information is
quickly gone if one does not save the file quickly enough.

Expanded the size to 30 MB, which is much more useful but still not
too big for modern file system standards.

Also fixed the setup od the BrowseRemoteProtocols directive when
creating cups-browsed.conf.

This change affects only new installations of the Snap, when updating,
after the update, either manually edit

    /var/snap/cups/common/etc/cups/cups-browsed.conf

to have a line

    DebugLogFileSize 30000

or delete

    /var/snap/cups/common/etc/cups/cups-browsed.conf

and in both cases, restart the CUPS Snap:

   sudo snap stop cups
   sudo snap start cups