Do not build libsnapd-glib from GIT Master any more
In the beginning of the CUPS Snap development we have added
functionality to snapctl and the equivalent library function
snapd_client_run_snapctl2_sync() to make Snap mediation in CUPS
working, letting cupsd block administrative inquiries if they are from
a confined Snap which does not plug the "cups-control" interface. To
make use of these changes immediately we needed to build libsnapd-glib
from GIT Master instead of using a released version.

libsnapd-glib from GIT Master kept working correctly all the time so
that this aspect got simply forgotten.

But know the Snap mediation ceased to work and as a wild guess to fix
this I removed the "libsnapd-glib" part and installed the released
version via build-packages and snap-packages in the "cups" part, which
solved the problem.

The CUPS Snap is still using core20 and it seems that the latest and
greatest libsnapd-glib does not work well with core 20 any more.