Use slot name for snapctl-based Snap mediation by cupsd
When CUPS is snapped it does the Snap mediation (allow administrative
inquiries from client Snaps only if they plug cups-control) via
"snapctl", or exactly the equivalent API function of libsnapd-glib,
"snapd_client_run_snapctl2_sync()". This API function does not check
for the name of the interface (or the client's plug), "cups-control"
but for the CUPS Snap's slot, named "cups-server". This broke when we
renamed the "cups-control" slot to "cups-server" in commit ced425ae2
and overlooked to test the standalone mode (we were worrying about the
proxy mode, and we do not have CI yet).

Corrected CUPS to use said API function with "cups-server" now.