Rename the plugs and slots for easier use of "cups-control"
- Renamed the "cups-server" slot to "cups-control", to be plugged as
  "cups-cups-control" by the "cups-control" plugs of applications.

- Renamed "cups-control" plug to "cups-host", to avoid name clashes
  by the renaming of the "cups-server" slot.

- Let the included command line tools plug "cups-control" via the
  "cups-host" plug.

This naming should be more intuitive (user plaggging applications's "cups-control" to CUPS' "cups:cups-control" and CUPS' plug "cups-host" is not actiually handled by end users. Also it could be easier to implement fallbacks in cupsd, like use `cups:cups-control` if `:cups-control` is missing or not backed by a CUPS daemon.