Added "install-cups-runtime-dependency" slot for pseudo content interface
This is a slot for the pseudo content interface for clients which plug
the "cups" interface and need to get the CUPS Snap auto-installed.

Actually, the pseudo content interface already works without the CUPS
Snap providing a slot for it, but providing the slot prevents an error
when an Ubuntu Core image with a client application plugging "cups"
and the CUPS Snap is built (using ubuntu-image).